Split group result insight parsing
