+++ apps/web/src/features/recommendation/groupResultInsights.ts
@@ -1,4 +1,13 @@
-type UnknownRecord = Record<string, unknown>;
+import {
+  cleanString,
+  cleanStringArray,
+  cleanStringValues,
+  groupPeople,
+  hasFavoriteActor,
+  intersectMoodPreferences,
+  quizPeople,
+  uniqueStrings,
+} from './groupResultInsightInput';
 
 export interface GroupResultInsights {
   participantNames: string[];
@@ -22,60 +31,12 @@ export function getQuizPeopleCount(quizData: unknown): number {
 }
 
 export function hasFavoriteActorSignal(quizData: unknown): boolean {
-  const people = Array.isArray(quizData) ? quizData : [quizData];
-  return people.some((person) => isRecord(person) && cleanString(person.favoriteActor) !== null);
-}
-
-function isRecord(value: unknown): value is UnknownRecord {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function cleanString(value: unknown): string | null {
-  if (typeof value !== 'string') return null;
-  const trimmed = value.trim();
-  return trimmed.length > 0 ? trimmed : null;
-}
-
-function cleanStringArray(value: unknown): string[] {
-  if (!Array.isArray(value)) return [];
-  return value.map(cleanString).filter((item): item is string => item !== null);
-}
-
-function cleanStringValues(values: unknown[]): string[] {
-  return values.map(cleanString).filter((item): item is string => item !== null);
-}
-
-function uniqueStrings(values: string[]): string[] {
-  const seen = new Set<string>();
-  const result: string[] = [];
-
-  for (const value of values) {
-    const key = value.toLocaleLowerCase();
-    if (seen.has(key)) continue;
-    seen.add(key);
-    result.push(value);
-  }
-
-  return result;
-}
-
-function intersectMoodPreferences(people: UnknownRecord[]): string[] {
-  const moodGroups = people.map((person) => cleanStringArray(person.moodPreference));
-  if (moodGroups.some((moods) => moods.length === 0)) return [];
-
-  const [firstGroup = []] = moodGroups;
-  return firstGroup.filter((mood) =>
-    moodGroups.every((moods) =>
-      moods.some((candidate) => candidate.toLocaleLowerCase() === mood.toLocaleLowerCase()),
-    ),
-  );
+  return quizPeople(quizData).some(hasFavoriteActor);
 }
 
 export function buildGroupResultInsights(quizData: unknown): GroupResultInsights | null {
-  if (!Array.isArray(quizData) || quizData.length < 2) return null;
-
-  const people = quizData.filter(isRecord);
-  if (people.length < 2) return null;
+  const people = groupPeople(quizData);
+  if (!people) return null;
 
   return {
     participantNames: people.map(
